server/pyproject.toml
[tool.poetry]
name = "Handigo Service"
version = "0.1.0"
description = ""
authors = ["Greg Ezema <gregobinna@gmail.com>"]
readme = "README.md"
packages = [
    { from = "src", include = "handigo_service" }
]


[tool.poetry.scripts]
handigo-cli = "handigo_service.interface.cli.app:main"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('.*'\\):DeprecationWarning",
    "ignore::DeprecationWarning:google.rpc",
]


[tool.poetry.dependencies]
python = "^3.12.1"
requests = "^2.31.0"
statsd = "^3.3.0"
sqlmodel = "^0.0.22"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
environs = "^9.5.0"
orjson = "^3.9.2"
sqlalchemy-utils = "^0.41.1"
click = "^8.1.6"
fastapi = "^0.111.0"
uvicorn = {version = "^0.24.0", extras = ["standard"]}
asgiref = "^3.7.2"
asyncpg = "^0.29.0"
dataclasses-json = "^0.6.7"
retry = "^0.9.2"
aiohttp = "^3.12.15"
redis = "^6.0.0"
passlib = "^1.7.4"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
ruff = "^0.1.13"
types-requests = "^2.31.0.1"
pytest = "^7.4.0"
pytest-responses = "^0.5.1"
freezegun = "^1.2.2"
bandit = "^1.7.5"
ipdb = "^0.13.13"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
locust = "^2.24.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "statsd"
ignore_missing_imports = true

[tool.coverage]
    [tool.coverage.report]
    skip_empty = true

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
